# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Invoice App API
  version: "0.9.0"
  description: |
    Original API specification for a new invoicing web app. It is INSPIRED BY
    invoice-generator.com (same idea: fill a form, pick a document type, get a
    professional PDF) but is a separate, independent product with its own
    backend, its own data model, and its own UI — nothing here calls or depends
    on invoice-generator.com.

    Scope of this version (Phase 1, as agreed with the product owner):
      1. User accounts with full login (JWT), not bare API keys.
      2. A "document types" listing so the main page can show the user a choice
         of what to generate (invoice, receipt, quote, ...).
      3. Creating and downloading documents (PDF) by filling out a form.
      4. Because documents belong to a logged-in user, they are also listed here
         as a lightweight "history" (list/get/delete) — this falls out naturally
         from having accounts, even though it wasn't explicitly requested.

    Explicitly OUT of scope for this version (left for a later phase — see the
    x-future-phases block at the bottom of this file): payment collection/
    webhooks. (Customers, as of 0.3.0, and emailing documents to customers, as
    of 0.8.0, are no longer deferred — see below.)

    Changelog:
      0.2.0 — Resolved three open design questions from the spec review:
        - Money fields now use the reusable `MonetaryAmount` schema and are
...
security:
  - BearerAuth: []
...
servers:
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Create a new account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          description: An account with this email already exists.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Log in with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: >-
            Incorrect email or password, OR the account is temporarily locked
            after too many failed attempts (see x-security-policy.account_lockout —
            a lock currently surfaces as 401, indistinguishable from bad credentials).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Exchange a refresh token for a new access token
      description: |
        Access tokens are short-lived (see `expires_in` on AuthResponse).
        Call this before/when one expires, using the longer-lived refresh
        token, instead of asking the user to log in again.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: New token pair issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          description: Refresh token is invalid, expired, or was revoked (e.g. by logout).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/logout:
    post:
      operationId: logoutUser
      tags: [auth]
      summary: Revoke the current refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '204':
          description: Logged out — the refresh token no longer works.
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the logged-in user's profile
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
    patch:
      operationId: updateCurrentUser
      tags: [auth]
      summary: Update the logged-in user's profile (e.g. default business info)
      description: |
        Lets a user save their own business name/address/logo once, so it can
        pre-fill the "Who is this from?" field on the document form instead of
        retyping it every time.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/me/logo:
    post:
      operationId: uploadUserLogo
      tags: [auth]
      summary: Upload a business logo from the user's device
      description: |
        Stores the image server-side (Postgres) and sets the user's `logo_url` to a
        short, browser-loadable retrieval URL (`GET /auth/logo/{logoId}`), returning the
        updated profile so the frontend can drop it straight into its auth state.

        The server is the authority on the image: it sniffs the actual bytes (the declared
        content type is not trusted), rejects anything it can't decode as a raster image
        (PNG/JPEG/WebP/GIF; SVG is rejected), rejects uploads over 5 MB, and downscales +
        re-encodes to a bounded PNG. Each upload gets a fresh `logoId`, so the retrieval
        URL changes and browser/PDF caches can't mask a re-upload.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: The image file (max 5 MB; PNG/JPEG/WebP/GIF).
      responses:
        '200':
          description: Logo stored; returns the updated profile with the new `logo_url`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          description: >-
            No file, an unreadable/non-image file, or a file over 5 MB.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '429':
          $ref: '#/components/responses/TooManyRequests'
    delete:
      operationId: deleteUserLogo
      tags: [auth]
      summary: Clear the user's uploaded logo
      description: >-
        Removes the stored image and nulls `logo_url`, returning the updated profile.
        (Setting `logo_url` to `""` via `PATCH /auth/me` also clears the URL, but only this
        endpoint removes the stored bytes.)
      responses:
        '200':
          description: Logo cleared; returns the updated profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/logo/{logoId}:
    parameters:
      - name: logoId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: getUserLogo
      tags: [auth]
      summary: Fetch a stored logo image by its opaque id
      description: >-
        Public and unauthenticated — an `<img>` tag (and the PDF renderer) can't send a
        bearer token, so access control is the unguessable, non-enumerable `logoId` rather
        than the user id. Not covered by the `/auth` rate limiter.
      security: []
      responses:
        '200':
          description: The logo image bytes.
          content:
            image/png:
              schema: { type: string, format: binary }
        '404':
          description: No logo exists with that id.

  # ------------------------------------------------------- DOCUMENT TYPES ----
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Access token returned by `/auth/login`, `/auth/register`, or
        `/auth/refresh`. Send as header "Authorization" with value
        "Bearer <access_token>".
...
  responses:
    Unauthorized:
      description: Missing, invalid, or expired access token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    ValidationError:
      description: The request body failed validation.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ValidationErrorResponse' }
    TooManyRequests:
      description: >-
        Rate limit exceeded (see x-security-policy.rate_limiting). Retry after
        the window resets.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: "invalid_credentials"
            message:
              type: string
              example: "Incorrect email or password."
    ValidationErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, fields]
          properties:
            code: { type: string, example: "validation_failed" }
            message: { type: string, example: "One or more fields are invalid." }
            fields:
              type: array
              items:
                type: object
                properties:
                  field: { type: string, example: "items[0].unit_cost" }
                  message: { type: string, example: "must be a positive number" }
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password, minLength: 8 }
        business_name: { type: string }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
    UpdateUserRequest:
      type: object
      properties:
        business_name: { type: string }
        business_address: { type: string, description: "Multiline, used to pre-fill the 'from' field." }
        logo_url: { type: string, format: uri }
        default_currency: { type: string, example: "EUR" }
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        business_name: { type: string, nullable: true }
        business_address: { type: string, nullable: true }
        logo_url: { type: string, format: uri, nullable: true }
        default_currency: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
    TokenPair:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        token_type: { type: string, example: "Bearer" }
        expires_in: { type: integer, description: "Access token lifetime in seconds.", example: 900 }
    AuthResponse:
      allOf:
        - $ref: '#/components/schemas/TokenPair'
        - type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
...